// playwrigh_api_tests/tests/jsonplaceholder/Schemas/users.schema.ts
import {z} from 'zod';


export const UserSchema = z.object({ // For one object
    id: z.number(),
    name: z.string(),
    username: z.string(),
    email: z.string().email(),
    address: z.object({
        street: z.string(),
        suite: z.string(),
        city: z.string(),
        zipcode: z.string(),
        geo: z.object({
            lat: z.string(),
            lng: z.string()
        })
    }),
    phone: z.string(),
    website: z.string(),
    company: z.object({
        name: z.string(),
        catchPhrase: z.string(),
        bs: z.string()
    })
});

export const UsersSchema = z.array(UserSchema); // For array of objects (note the different const names)